Make the action to upload both sarif and astgrep signals
Instead of running the linters within this action, delegate that part to
the caller and make this action responsible for uploading the results of
the linters. This way, the users can just pick whatever linter they want
to use and then use this action to upload the results to Aviator.